requirements.txt
asgiref==3.3.4
Django==3.2.4
django-multiselectfield==0.1.12
django-widget-tweaks==1.4.8
pytz==2021.1
sqlparse==0.4.1
